pyproject.toml
[tool.poetry]
name = "nlx"
version = "0.1.0"
description = "Python SDK for interacting with the NLx API and other affiliated tools"
authors = ["jjorissen52 <jjorissen52@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
nlx = "nlx.main:main"

[tool.poetry.dependencies]
python = "^3.7"
fire = "^0.4.0"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
rich = "^10.16.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "21.12b0"
isort = "^5.10.1"
pre-commit = "^2.16.0"
ipython = "^7.30.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = """
/(
  venv
  | \\.github
  | \\.git
)/
"""

[tool.isort]
line_length = 120
known_first_party = "nlx"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
multi_line_output = 3
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
